refactor: PostsFormat 코드 정리 및 리팩토링
