+++ app/(home)/(board)/general-market/page.tsx
@@ -1,7 +1,7 @@
-import PostsFormatWrapper from '@/components/Posts/PostsFormatWrapper';
+import PostFormat from '@/components/Posts/PostFormat';
 
 const generalMarketPage = () => {
-  return <PostsFormatWrapper />;
+  return <PostFormat />;
 };
 
 export default generalMarketPage;
