+++ app/(home)/globals.css
@@ -42,6 +42,7 @@
     background-color: #f9f9f9;
   }
 }
+
 @tailwind components;
 @tailwind utilities;
