+++ app/apis/images.ts
@@ -1,0 +1,1 @@
+export const getImageURL = () => {};
