+++ app/components/Board/index.tsx
@@ -1,16 +1,18 @@
 'use client';
-import { createContext, useContext, useEffect, useState } from 'react';
-import { DocumentData } from 'firebase/firestore';
+import { createContext, useContext, useEffect, useMemo, useState } from 'react';
+import { DocumentData, doc, updateDoc } from 'firebase/firestore';
 import { BoardProps, PostContextType } from './types';
 import { authState } from '@/recoil/atoms';
 import { useRecoilValue } from 'recoil';
 import { usePathname } from 'next/navigation';
 import getAdmin from '@/utils/getAdmin';
-import { Bodys } from './Bodys';
+import PostsLoading from '../Posts/PostsLoading';
+import PostsNotFound from '../Posts/PostsNotFound';
+import { dbService } from '@/firebase';
+import { deletePost } from '@/apis/posts';
+import PostList from './PostList';
 
 const defaultPostContext: PostContextType = {
-  posts: [],
-  editPosts: () => {},
   user: null,
   admin: null,
   pathname: '/',
@@ -20,16 +22,11 @@ export const PostContext = createContext(defaultPostContext);
 
 export const usePostContext = () => useContext(PostContext);
 
-const Board = ({ children, data }: BoardProps) => {
-  const [posts, setPosts] = useState<DocumentData[] | null>(data);
+const Board = ({ children }: BoardProps) => {
   const [admin, setAdmin] = useState<DocumentData | null>(null);
 
   const pathname = usePathname();
   const user = useRecoilValue(authState);
-
-  const editPosts = (newPosts: DocumentData[]) => {
-    setPosts(newPosts);
-  };
 
   useEffect(() => {
     if (!admin) {
@@ -42,7 +39,7 @@ const Board = ({ children, data }: BoardProps) => {
   }, [admin]);
 
   return (
-    <PostContext.Provider value={{ posts, editPosts, user, pathname, admin }}>
+    <PostContext.Provider value={{ user, pathname, admin }}>
       <ul className='w-full border-b border-neutral-500'>{children}</ul>
     </PostContext.Provider>
   );
@@ -69,6 +66,72 @@ const Headers = ({ type = 'etc' }: { type?: 'community' | 'etc' }) => {
       <div className='w-[6%] hidden lg:block'>조회수</div>
     </li>
   );
+};
+
+export const Bodys = ({
+  posts,
+  editPosts,
+  isLoading,
+  type = 'etc',
+}: {
+  posts: DocumentData[] | null;
+  editPosts: (value: DocumentData[]) => void;
+  isLoading: boolean;
+  type?: 'etc' | 'community';
+}) => {
+  const { user, pathname } = useContext(PostContext);
+
+  const handleDeletePost = (id: string) => {
+    const post = posts?.find((item) => item.id === id);
+    if (!post || !posts) return;
+
+    const ok = window.confirm('이 게시물을 삭제하시겠습니까?');
+
+    if (ok) {
+      deletePost(post, user, pathname, id);
+      const deletePosts = posts.filter((item) => item.id !== id);
+      editPosts(deletePosts);
+    }
+  };
+
+  const handleClickViewUp = async (id: string) => {
+    const post = posts && posts.find((item) => item.id === id);
+
+    if (post) {
+      const newPostObj = { ...post, views: post.views + 1 };
+      const docRef = doc(dbService, `${pathname}/${post.id}`);
+
+      await updateDoc(docRef, newPostObj);
+    }
+  };
+
+  const BodyContent = useMemo(() => {
+    !isLoading ? (
+      posts && posts?.length !== 0 ? (
+        posts.map((post) => {
+          return (
+            <PostList
+              key={post.id}
+              type={type}
+              pathname={pathname}
+              post={post}
+              user={user}
+              handleDeletePost={handleDeletePost}
+              handleClickViewUp={handleClickViewUp}
+            />
+          );
+        })
+      ) : (
+        /* 게시물 데이터가 없을 때 */
+        <PostsNotFound />
+      )
+    ) : (
+      /* 게시물 데이터 로딩 중 */
+      <PostsLoading />
+    );
+  }, []);
+
+  return <>{BodyContent}</>;
 };
 
 Board.Headers = Headers;
