+++ app/utils/getPosts.ts
@@ -27,13 +27,6 @@ const getPosts = async (pathname: string) => {
     return post.push({ id: doc.id, ...doc.data() });
   });
 
-  /* 
-  post.sort(function (a: DocumentData, b: DocumentData) {
-    return Number(b.createdAt) - Number(a.createdAt);
-  }); */
-
-  console.log('🚀 ~ getPosts ~ post:', post);
-
   if (post.length > 0) return post;
 
   return;
